+++ oracle-of-bacon-backend/src/main/java/com/serli/oracle/of/bacon/api/APIEndPoint.java
@@ -30,6 +30,10 @@ public class APIEndPoint {
 
     @Get("bacon-to?actor=:actorName")
     public String getConnectionsToKevinBacon(String actorName) {
+        try {
+            this.redisRepository.addSearch(java.net.URLDecoder.decode(actorName, "UTF-8"));
+        } catch (UnsupportedEncodingException ignored) {
+        }
 
         List<AbstractMap.SimpleEntry<String, Neo4JRepository.GraphItem>> result = this
                 .neo4JRepository
@@ -51,11 +55,7 @@ public class APIEndPoint {
 
     @Get("last-searches")
     public List<String> last10Searches() {
-        return Arrays.asList("Peckinpah, Sam",
-                "Robbins, Tim (I)",
-                "Freeman, Morgan (I)",
-                "De Niro, Robert",
-                "Pacino, Al (I)");
+        return redisRepository.getLastTenSearches();
     }
 
     @Get("actor?name=:actorName")