+++ oracle-of-bacon-backend/src/main/java/com/serli/oracle/of/bacon/repository/RedisRepository.java
@@ -6,13 +6,17 @@ import java.util.List;
 
 public class RedisRepository {
     private final Jedis jedis;
-
+    private final String searchesKey = "bacon_search";
     public RedisRepository() {
         this.jedis = new Jedis("localhost");
     }
 
     public List<String> getLastTenSearches() {
-        // TODO implement last 10 searchs
-        return null;
+        return jedis.lrange(searchesKey, 0, -1);
+    }
+
+    public void addSearch(String search) {
+        jedis.lpush(searchesKey, search);
+        jedis.ltrim(searchesKey, 0, 9);
     }
 }